requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
tkinter>=8.6
pyperclip>=1.8.2
quantumrandom>=1.9.0
tensorflow>=2.8.0
scikit-learn>=1.0.2
qiskit>=0.34.2
biopython>=1.79
torch>=1.10.0
